Refactor Label & Bar for DeployState
