+++ src/UI/Components/ResourceStatus/ColorConfig.ts
