--- src/UI/Pages/Resources/Components/TableControls/index.ts
@@ -1,1 +1,0 @@
-export * from "./ResourceTableControls";
