+++ src/UI/Pages/Resources/Components/index.ts
@@ -1,5 +1,3 @@
-export * from "./DeployStateColorConfig";
-export * from "./DeployStateBar";
-export * from "./TableControls";
+export * from "./ResourceTableControls";
 export * from "./DeployButton";
 export * from "./RepairButton";
