+++ src/UI/Pages/Resources/ResourceTableRow.tsx
@@ -1,12 +1,12 @@
 import React, { useContext } from "react";
 import { Link } from "react-router-dom";
-import { Button, Label } from "@patternfly/react-core";
+import { Button } from "@patternfly/react-core";
 import { InfoCircleIcon } from "@patternfly/react-icons";
 import { Tbody, Tr, Td } from "@patternfly/react-table";
 import { Resource } from "@/Core";
+import { ResourceStatusLabel } from "@/UI/Components";
 import { DependencyContext } from "@/UI/Dependency";
 import { words } from "@/UI/words";
-import { labelColorConfig } from "./Components";
 
 interface Props {
   row: Resource.Row;
@@ -24,9 +24,7 @@ export const ResourceTableRow: React.FC<Props> = ({ row }) => {
           {row.numberOfDependencies}
         </Td>
         <Td dataLabel={words("resources.column.deployState")}>
-          <Label color={labelColorConfig[row.deployState]}>
-            {row.deployState}
-          </Label>
+          <ResourceStatusLabel status={row.deployState} />
         </Td>
         <Td>
           <Link
